fibolimit

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -6,14 +6,14 @@
 int main(void)
 {
 	int a = 0;
-	unsigned long int x = 0, y = 1, next_value = 0;
+	int x = 0, y = 1, next_value = 0;
 
 	while (a < 98)
 	{
 		next_value = x + y;
 		x = y;
 		y = next_value;
-		printf("%lu", next_value);
+		printf("%d", next_value);
 		if (a < 97)
 			printf(", ");
 		a++;
